Introduce Zero constant to EulerAngles class; remove remaining old angle converter functions

+++ TR5Main/Scripting/GameScriptRotation.cpp
@@ -37,7 +37,7 @@ void GameScriptRotation::Register(sol::state* state)
 @return A Rotation object.
 @function Rotation.new
 */
-GameScriptRotation::GameScriptRotation(int aX, int aY, int aZ)
+GameScriptRotation::GameScriptRotation(float aX, float aY, float aZ)
 {
 	x = aX;
 	y = aY;
